requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
authlib
requests